requirements.txt
feedparser==6.0.11
html2text==2024.2.26
mdit_py_plugins==0.4.0
nltk==3.8.1
notion_client==2.2.1
openai==1.23.2
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
Requests==2.31.0
markdown-it-py
beautifulsoup4>=4.12.3
pytz